remove old postgres metrics
