+++ packages/fuel-indexer-database/postgres/Cargo.toml
@@ -13,11 +13,10 @@ description = "Fuel Indexer Postgres"
 fuel-indexer-database-types = { workspace = true }
 fuel-indexer-lib = { workspace = true }
 fuel-indexer-macro-utils = { workspace = true, optional = true }
-fuel-indexer-metrics = { workspace = true, optional = true }
 sqlx = { version = "0.6", features = ["runtime-tokio-rustls", "postgres", "offline"] }
 tracing = "0.1"
 uuid = { version = "1.3", features = ["v4"] }
 
 [features]
 default = ["metrics"]
-metrics = ["fuel-indexer-metrics", "fuel-indexer-macro-utils"]
+metrics = ["fuel-indexer-macro-utils"]
