--- packages/fuel-indexer-metrics/src/db/mod.rs
@@ -1,2 +1,0 @@
-mod postgres;
-pub use postgres::Postgres;
